cmd/participant: clean up client
