+++ cmd/participant/client.go
@@ -5,7 +5,6 @@ import (
 	"encoding/json"
 	"errors"
 	"fmt"
-	"io"
 	"io/ioutil"
 	"net/http"
 	"time"
@@ -14,9 +13,8 @@ import (
 )
 
 type Client struct {
-	url          string
-	registration *registration
-	sessionID    string
+	url       string
+	sessionID string
 }
 
 func NewClient(url string) *Client {
@@ -25,53 +23,8 @@ func NewClient(url string) *Client {
 	}
 }
 
-type registration struct {
-	Start    int
-	Deadline int
-	Ticket   string
-}
-
-func (c *Client) StartTime() *time.Time {
-	if c.registration == nil {
-		return nil
-	}
-	t := time.Unix(int64(c.registration.Start), 0)
-	return &t
-}
-
-func (c *Client) Register() error {
-	url := fmt.Sprintf("%v/%v", c.url, "participation")
-	var body io.Reader
-	resp, err := http.Post(url, "text/html; charset=UTF-8", body)
-	if err != nil {
-		return err
-	}
-	responseData, err := ioutil.ReadAll(resp.Body)
-	if err != nil {
-		return err
-	}
-	fmt.Println(string(responseData))
-	var part registration
-	if err := json.Unmarshal(responseData, &part); err != nil {
-		return err
-	}
-	c.registration = &part
-	fmt.Printf("Registered for ceremony at time %v\n", time.Unix(int64(part.Start), 0))
-	return nil
-}
-
-type Info struct {
-	Start    int
-	Deadline int
-	Ceremony *towersofpau.JSONCeremony
-}
-
-func (c *Client) GetCeremony() (*Info, error) {
-	fmt.Println("Fetching Ceremony")
-	if c.registration == nil {
-		return nil, errors.New("no registration available")
-	}
-	url := fmt.Sprintf("%v/%v/%v", c.url, "participation", c.registration.Ticket)
+func (c *Client) GetStatus() (*CeremonyStatus, error) {
+	url := fmt.Sprintf("%v/%v/%v", c.url, "info", "status")
 	resp, err := http.Get(url)
 	if err != nil {
 		return nil, err
@@ -80,39 +33,130 @@ func (c *Client) GetCeremony() (*Info, error) {
 	if err != nil {
 		return nil, err
 	}
-	var info Info
+	var info CeremonyStatus
 	if err := json.Unmarshal(responseData, &info); err != nil {
 		return nil, err
 	}
 
-	c.registration.Start = info.Start
-	c.registration.Deadline = info.Deadline
-
-	fmt.Println("Retrieved ceremony")
+	fmt.Printf("Ceremony status: %v\n", info)
 	return &info, nil
 }
 
-func (c *Client) SubmitCeremony(ceremony *towersofpau.Ceremony) error {
-	fmt.Println("Submitting ceremony")
-	if c.registration == nil {
-		return errors.New("no registration available")
+func (c *Client) GetCurrentState() (*towersofpau.Ceremony, error) {
+	url := fmt.Sprintf("%v/%v/%v", c.url, "info", "current_state")
+	fmt.Printf("Getting current state from %v\n", url)
+	resp, err := http.Get(url)
+	if err != nil {
+		return nil, err
 	}
-	url := fmt.Sprintf("%v/%v/%v", c.url, "participation", c.registration.Ticket)
+	responseData, err := ioutil.ReadAll(resp.Body)
+	if err != nil {
+		return nil, err
+	}
+	start := time.Now()
+	defer func() {
+		fmt.Printf("Deserializing took %v\n", time.Since(start))
+	}()
+	return towersofpau.Deserialize(bytes.NewReader(responseData))
+}
+
+func (c *Client) Contribute(contribution *towersofpau.BatchContribution) error {
+	fmt.Printf("Submitting our contribution\n")
+	url := fmt.Sprintf("%v/%v", c.url, "contribute")
+	marshalled, err := json.Marshal(contribution)
+	if err != nil {
+		return err
+	}
+
+	request, err := http.NewRequest("POST", url, bytes.NewBuffer(marshalled))
+	if err != nil {
+		return err
+	}
+	request.Header.Set("Content-Type", "application/json")
+	request.Header.Add("Authorization", "Bearer "+c.sessionID)
+	resp, err := http.DefaultClient.Do(request)
+	if err != nil {
+		return err
+	}
+
+	responseData, err := ioutil.ReadAll(resp.Body)
+	if err != nil {
+		return err
+	}
+	fmt.Printf("Response: %v\n", string(responseData))
+
+	if err := handleStatus(resp.StatusCode); err != nil {
+		return err
+	}
+	return nil
+}
+
+func (c *Client) Abort(p *Participant) error {
+	url := fmt.Sprintf("%v/%v/%v", c.url, "contribution", "abort")
 
 	buf := new(bytes.Buffer)
-	towersofpau.Serialize(buf, ceremony)
+	encoder := json.NewEncoder(buf)
+	if err := encoder.Encode(p); err != nil {
+		return err
+	}
 	resp, err := http.Post(url, "text/html", buf)
 	if err != nil {
 		return err
 	}
-	switch resp.StatusCode {
+	if err := handleStatus(resp.StatusCode); err != nil {
+		responseData, err := ioutil.ReadAll(resp.Body)
+		if err != nil {
+			return err
+		}
+		fmt.Printf("Response: %v\n", string(responseData))
+		return err
+	}
+	return nil
+}
+
+func (c *Client) TryContribute() (*towersofpau.BatchContribution, error) {
+	url := fmt.Sprintf("%v/%v/%v", c.url, "lobby", "try_contribute")
+	fmt.Printf("Trying to contribute at %v with ssid %v\n", url, c.sessionID)
+
+	request, err := http.NewRequest("POST", url, nil)
+	if err != nil {
+		return nil, err
+	}
+	request.Header.Add("Authorization", "Bearer "+c.sessionID)
+	resp, err := http.DefaultClient.Do(request)
+	if err != nil {
+		return nil, err
+	}
+	responseData, err := ioutil.ReadAll(resp.Body)
+	if err != nil {
+		return nil, err
+	}
+
+	var errorUnm ErrorStruct
+	if err := json.Unmarshal(responseData, &errorUnm); err == nil && len(errorUnm.Error) > 0 {
+		return nil, fmt.Errorf("%v: %v", errorUnm.Code, errorUnm.Error)
+	}
+
+	if err := handleStatus(resp.StatusCode); err != nil {
+		fmt.Printf("Response: %v", string(responseData))
+		return nil, err
+	}
+
+	var contribution towersofpau.BatchContribution
+	if err := json.Unmarshal(responseData, &contribution); err != nil {
+		return nil, err
+	}
+	return &contribution, nil
+}
+
+func handleStatus(status int) error {
+	switch status {
 	case 200:
-		fmt.Println("Submitted ceremony successfully")
 		return nil
 	case 400:
 		return errors.New("invalid ceremony")
 	case 403:
 		return errors.New("invalid ticket provided")
 	}
-	return errors.New("invalid status code")
+	return fmt.Errorf("invalid status code: %v", status)
 }
